requirements.txt
torch
faiss-cpu
transformers
pdfplumber
